Fix regression with Swagger 1.2 specs - #43
