--- tests/common/validate_json_test.py
@@ -1,17 +1,0 @@
-import json
-import os
-import pytest
-
-from swagger_spec_validator.common import SwaggerValidationError, validate_json
-
-
-def test_success():
-    my_dir = os.path.abspath(os.path.dirname(__file__))
-    with open(os.path.join(my_dir, '../data/v2.0/petstore.json')) as petstore_file:
-        petstore_spec = json.load(petstore_file)
-    validate_json(petstore_spec, 'schemas/v2.0/schema.json')
-
-
-def test_failure():
-    with pytest.raises(SwaggerValidationError):
-        validate_json({}, 'schemas/v2.0/schema.json')
